Rename hazard and validation python codes
Rename hazard and validation python codes
